[threadpool-ms] Exit worker thread after 5 to 60 seconds
The timeout before exiting a worker thread is random to limit the worst
case scenario where, if the timeout is set to N seconds, the user would
launch a lot of work item simultaneously, at an interval of N + 1
seconds. In this case, we would have to kill all the threads at once,
and starts them all over again 1 second later. The randomization makes
the creation and destruction more linear in time, and avoid burst.