[MSBuild] Optimize build time by caching build results
xbuild uses MSBuild tasks to resolve project references and
for getting the list of files to copy from referenced projects
(by invoking the targets GetTargetPath and
GetCopyToOutputDirectoryItems).

In solutions with many projects and many references, this may be
very slow, since those targets have to be invoked recursively
through all references.

This patch adds a cache for the results of those targets, so that
the target is invoked only once per build session.