[threads] Avoid using same signals as sgen-os-posix.c
This would trigger a hang with finalizer-threads.cs, because sgen-stw
would send a signal, but it would be the mono-threads-posix-signal.c
suspend_signal_handler which would be run for the signal. That would
trigger a deadlock, because sgen_wait_for_suspend_ack would wait
indefinitely for the suspend_ack_semaphore_ptr which would never be
posted.