[acceptance-tests] Pull in new CoreCLR tests added upstream
A bunch of tests were added in the upstream repo, add those to the source listings.

There's a suite of tests in the CoreMangLib folder which are mostly testing corelib,
added a separate target "coreclr-runtest-coremanglib" for it and renamed the existing
target to "coreclr-runtest-basic".

Build and link all cs testcases with the coreclr-testlibrary.dll which enables a
few more tests to run.