[suspend] Fix suspend race between mono-threads and sgen suspend when using the posix backend.
The problem happens when a thread is left suspended by mono-threads and sgen tries to suspend it.

The signal mask of sigsuspend call in mono-threads didn't include sgen's suspend signal, making
it impossible for sgen to suspend it.

The fix is to include sgen's suspend signal into mono-threads sigsuspend mask. This way sgen
suspend can nest over mono-thread's.

The opposite is not needed as sgen never leaves threads suspended when it finishes and it holds
the suspend lock the whole time, making it impossible for mono-threads to initiate any suspend operation.