[acceptance-tests] Split compiling/cleaning test sources and bump coreclr
The coreclr-compile-tests would throw a shell error "Argument list too long" on OSX due
to the huge amount of tests passed to make in one go. The same happened for make clean.

We now compile and clean the three sets of coreclr test sources separately to avoid this.

Additionally, bump the coreclr reference and include a new test that was added upstream.