CSharpCodeCompiler should not treat multi-line warnings from mcs as internal compiler errors.
Most errors and warnings from c# compiler are reported in one line in format "FILE(LINE, COL): warning CS1234: Text".
But some warning span multiple lines:

  /tmp/3fb23cc0/634d2022.0.cs(22,29): warning CS0114: `NS.B.F()' hides inherited member `NS.A.F()' ..
  /tmp/3fb23cc0/634d2022.0.cs(16,29): (Location of the symbol related to previous warning)

Previously the second line was considered to be an internal compiler error since it doesn't match the error regex.

This seems to be only type of messages that span multiple lines.

This fixes #35980