Clarify ownership of generic param objects
This patch fixes rule violations found by the first batch of
mempool-reference audit points. The fix is to remove ambiguities in
what image owns or should own any particular generic parameter object.
Previously, there were four options for specifying a
MonoGenericParam’s image:

 1. klass/method union in MonoGenericContainer.
 2. image field in MonoGenericContainer.
 3. image field in MonoGenericParam.
 4. No field; “image” passed in to methods in at time of usage.

This patch removes 3 and 4; it preserves the image field from (2), but
merges it into the union from (1). It adds a new field,
“is_anonymous”, to indicate when the image member of the union should
be used. The single source of the “owner" union now always points to
the generic param’s image, directly or indirectly.

A further change is made to support this: Previously, containers were
optional, and the presence or absence of container was incidentally
used to determine whether an object was a MonoGenericParam or a
MonoGenericParamFull. Now, containers are mandatory for all
MonoGenericParams, so a new “is_small_size” bit in the container
tracks this information.

This change opens some sources of risk:

 ⁃ Code which creates MonoGenericParams now must make sure to
allocate a container early.
 ⁃ Code that accesses the MonoGenericContainer::owner field
must all be correctly updated to look for is_anonymous or there will
be a crash. (Much of this code would have previously used a NULL
container as the flag for “anonymous”).
 ⁃ Code that casts MonoGenericParams to MonoGenericParamFull
must use the is_small_size field.
 ⁃ Some generic containers will now exist with NULL type_params
and 0 type_argc. Code which reads the type param array must either not
receive objects from these containers, or must handle it.
 ⁃ If code by users of the embedding API exists which somehow
depends on particular behavior related to the now-unused image
argument of mono_class_from_generic_parameter, it will break.

I have audited code around, at least, the first three of these cases
and fixed what problems I found.