Remove an AOT-unfriendly code path.
This code path is intended as an optimization, but my tests have shown that it does not add significant performance.
Instead, it causes problems for AOT (with both Mono and IL2CPP) when the type of the property is a value type.
Since this code path already does not exist in the .NET Subset profile used by Unity, I'll also remove it from the full profile.