[amd64] Add is_gsharedvt_return_value to ArgInfo. Save to the stack ArgGSharedVtInReg args.
The usage of is_gsharedvt_return_value is easier than having a separate storage as supporting
it would require changing all usage sites and this is only needed by start_call.

All incoming arguments in registers must be saved on the stack as the rest of the JIT
expects to be able to load them from their well known places. Without this change, it would
load a random value instead.