[OSX] Use syscall for mmap and munmap on OSX. Fixes #19379
The problem with #19379 is that instruments interpose mmap and munmap
with code that takes a lock and record arguments in a global table.

It doesn't go without saying that taking locks is not signal safe
meaning it deadlocks when using async suspend.

The first attempt at fixing this was using a suspend critical section,
but that didn't work as that lock is taken for malloc/free too, which
are called by the whole world outside of the runtime too.

The alternative is to completely bipass instruments here and ensure the
lock is not taken. This requires us doing the syscall directly.

The only concern with this is OSX sandboxed apps, to ensure this is not a problem,
I added a configure option that will disable it for XM to use.