Compute the proper managed size for a packed hierarchy
This change corrects another issue related to the changes on
marshal-struct-hierachy-fix, merged to unity-staging at
6605fdac06b4a50ace3dc24f4d659ce7c511d28f.

We can't quite assume that the first field in a child is aligned with
the minimum alignment for the parent. Instead, we'll let the child tell
the parent about the offset of its first field, and use that.

For example, given this code:

[StructLayout(LayoutKind.Sequential, Pack = 4)]
public class Base
{
public ushort a;
public ushort b;
}

[StructLayout(LayoutKind.Sequential, Pack = 4, CharSet = CharSet.Ansi)]
public class Person : Base
{
public uint unType;
[MarshalAs(UnmanagedType.ByValTStr, SizeConst = 128)]
public string Name;
    public uint foo;
   }

The minimum alignment of Person is 8 on 64-bit (since the Name field is
a pointer). But the offset of the uType field in Person is at 4, since
that still meets the minimum alignment requirements. So the first field
of Person has an offset of 4 in the managed representation of the
structure. The previous code here assumed that offset is 8.