[handle] Add arena marks to support mono_raise_exception with coop handles.
The idea is that when we enter an icall from managed to native,
we put an unwind mark on the current top of the arena stack.
If we return normally, we clear the mark.
If an exception occurs, we unwind the stack until the current mark.

That way, handles that were used before the exception occurred are
cleaned up before we begin propagating the exception.

The easiest way to make use of unwind marks is using the
MONO_ICALL_ENTER () and MONO_ICALL_LEAVE () macros:

void
my_icall (...)
{
	MONO_ICALL_ENTER ();

	MONO_HANDLE_ARENA_PUSH ();
	...
	MONO_HANDLE_ARENA_POP ();

	MONO_ICALL_LEAVE ();
}