[handle] Remove MonoHandleArena chunk_last field.
This changes MonoHandleArena from a queue of chunks to a stack of
chunks.  More importantly, the MonoHandleArena has an initial chunk that
is allocated on the stack together with the arena.  Previously that
initial chunk was at the front of the list of chunks, but now it is at
the end.