[delegate] Refactor how AsyncResult are created for BeginInvoke
When doing Delegate.BeginInvoke, we return an AsyncResult, and it has to cover 2 possible cases:
 - the remoting case
 - the threadpool case (or non-remote case)

The code paths both starts at mono_delegate_begin_invoke. In the threadpool case, we call mono_threadpool_ms_begin_invoke.

The previous implementation of mono_async_result_new would take the following 5 parameters:
 - domain
 - handle: it would always be NULL
 - state: used in both cases
 - data: it would always be NULL
 - object_data: it would only be non-NULL in the threadpool case

As 3 out of these 5 parameters are either not used, or used in only in one of the 2 cases, it's easier to get rid of them.

They have been replaces with the following parameters:
 - domain
 - delegate: the delegate to asynchronously call
 - state: the state to pass the above delegate
 - callback: the callback to call after the delegate

As for object_data, it is now set in the threadpool case after having created the AsyncResult object.

Finally, as the value returned by AsyncResult.Invoke is never used (and can never be used by the threadpool), we can simply set its return type to void.